requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
msgpack==1.0.8
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
tarantool==1.1.2
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4

openpyxl~=3.1.2